refactor: Split up limit rendering files
