+++ app/charts/area/chart-area.tsx
@@ -15,7 +15,7 @@ import {
 import { Ruler } from "@/charts/shared/interaction/ruler";
 import { Tooltip } from "@/charts/shared/interaction/tooltip";
 import { LegendColor } from "@/charts/shared/legend-color";
-import { VerticalLimits } from "@/charts/shared/limits";
+import { VerticalLimits } from "@/charts/shared/limits/vertical";
 import { InteractionHorizontal } from "@/charts/shared/overlay-horizontal";
 import { AreaConfig } from "@/config-types";
 import { useLimits } from "@/config-utils";
